[#27693] RiscV: Fix comparison in vmax.vx

+++ arch/riscv/vector_helper_template.h
@@ -1706,22 +1706,22 @@ void glue(helper_vmax_ivi, POSTFIX)(CPUState *env, uint32_t vd, int32_t vs2, tar
         switch (eew) {
         case 8: {
                 int8_t v2 = ((int8_t *)V(vs2))[ei];
-                ((int8_t *)V(vd))[ei] = v2 > rs1 ? v2 : rs1;
+                ((int8_t *)V(vd))[ei] = v2 > (int8_t)rs1 ? v2 : (int8_t)rs1;
                 break;
             }
         case 16: {
                 int16_t v2 = ((int16_t *)V(vs2))[ei];
-                ((int16_t *)V(vd))[ei] = v2 > rs1 ? v2 : rs1;
+                ((int16_t *)V(vd))[ei] = v2 > (int16_t)rs1 ? v2 : (int16_t)rs1;
                 break;
             }
         case 32: {
                 int32_t v2 = ((int32_t *)V(vs2))[ei];
-                ((int32_t *)V(vd))[ei] = v2 > rs1 ? v2 : rs1;
+                ((int32_t *)V(vd))[ei] = v2 > (int32_t)rs1 ? v2 : (int32_t)rs1;
                 break;
             }
         case 64: {
                 int64_t v2 = ((int64_t *)V(vs2))[ei];
-                ((int64_t *)V(vd))[ei] = v2 > rs1 ? v2 : rs1;
+                ((int64_t *)V(vd))[ei] = v2 > (int64_t)rs1 ? v2 : (int64_t)rs1;
                 break;
             }
         default:
